ref: :recycle: renames item.ts to content.ts
